% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_over.R
\name{heatmap_over}
\alias{heatmap_over}
\title{A filtered ComplexHeatmap}
\usage{
heatmap_over(x, cutoff = 0)
}
\arguments{
\item{x}{Matrix of gene expression of x genes}

\item{cutoff}{double() Value to filter the matrix}
}
\value{
A figure of ComplexHeatmap::Heatmap type
}
\description{
This function allows to filter a matrix of gene expression and plot a ComplexHeatmap.
Values below a cutoof will be replaced by NA and plotted in gray
}
\examples{
# Creates a dummy gene expression matrix
expresion_genes <- matrix(rnorm(100), nrow = 10)
colnames(expresion_genes) <- paste0("especie_",letters[1:10])
rownames(expresion_genes) <- paste0("gene_",letters[1:10])

# Run example
subset_heatmap(expresion_genes)

}
