% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_heatmap.R
\name{subset_heatmap}
\alias{subset_heatmap}
\title{A ComplexHeatmap subseted by gen names}
\usage{
subset_heatmap(x, especies = NULL, genes = NULL)
}
\arguments{
\item{x}{matrix() expression level by genes (rows) and species (columns)}

\item{especies}{character() species to be selected in this function}

\item{genes}{character() genes to be selected in this function}
}
\value{
A plot of ComplexHeatmap::Heatmap type
}
\description{
Gene names are used to filter
}
\examples{
# Creates a dummy gene expression matrix
expresion_genes <- matrix(rnorm(100), nrow = 10)
colnames(expresion_genes) <- paste0("especie_",letters[1:10])
rownames(expresion_genes) <- paste0("gene_",letters[1:10])

# Run example
subset_heatmap(expresion_genes,
       especies = c("especie_a", "especie_b","especie_c"),
       genes = c("gene_d","gene_e","gene_f"))


}
